refactor(settlement): split domain modules and pin backend commands
