+++ .cspell.json
@@ -3,6 +3,7 @@
     "language": "en",
     "words": [
         "footgun",
+        "Finalizable",
         "KASHI",
         "Masaki",
         "MUSUBI",
