+++ apps/backend/docs/settlement_domain_types.md
@@ -10,12 +10,15 @@ The `musubi_settlement_domain` crate now owns:
 - canonical settlement identifiers such as `PaymentReceiptId`, `SettlementCaseId`, `SettlementIntentId`, `SettlementSubmissionId`, and `ObservationId`
 - distinct idempotency types: `InternalIdempotencyKey` and `ProviderIdempotencyKey`
 - safe money primitives: `CurrencyCode` and `Money`
-- backend descriptor and capability types
-- an opaque `ProviderPayload` type for adapter-specific input without pulling serde into the domain crate
+- backend descriptor, backend pinning, and capability types
+- typed provider payload schema/field/value types without pulling serde into the domain crate
 - provider-agnostic command, result, observation, and error enums
+- settlement state vocabulary: primary phase, resolution kind, overlays, and `SettlementState`
 - the pure `SettlementBackend` trait
 
 `SettlementIntentId` is used instead of `PromiseIntentId` because the pinned foundation documents define **Settlement Intent** as the canonical settlement object.
+
+The crate is also split into focused modules so the core contract is readable without one giant `lib.rs`.
 
 ## Why money is safe
 
@@ -39,6 +42,39 @@ There is no unchecked float-based arithmetic and no silent rescaling.
 
 They are intentionally different types so retries cannot accidentally reuse internal identifiers as if they were provider-safe wire values.
 The trait exposes provider-key derivation explicitly through `provider_idempotency_key(...)`.
+
+## Why backend pinning is explicit
+
+`BackendPin` carries the pinned `backend_key` and `backend_version` on command objects.
+This makes durable replay safer because orchestration can preserve which backend/version a case was pinned to when the command was created.
+
+`SettlementBackend::ensure_backend_pin(...)` fails closed if a command is routed to a backend descriptor that does not match the pinned backend.
+
+## Why provider payload is no longer bytes
+
+The first draft used `ProviderPayload(Vec<u8>)`.
+That shape was too permissive and invited future drift toward "just serialize anything."
+
+The payload is now:
+- schema-aware
+- field-based
+- typed through `ProviderPayloadValue`
+- still provider-agnostic
+
+This keeps the domain crate free from serde/JSON convenience leakage while still making payload structure explicit.
+
+## Settlement state model
+
+The foundation docs say settlement state is:
+- one primary phase
+- an optional resolution kind
+- zero or more overlays
+
+The crate now reflects that directly with:
+- `SettlementPrimaryPhase`
+- `SettlementResolutionKind`
+- `SettlementOverlay`
+- `SettlementState`
 
 ## What the trait does
 
@@ -64,5 +100,6 @@ This issue does not implement:
 - provider-specific adapters
 - database persistence
 - happy-route case progression
+- adapter-owned provider payload builders or registries
 
 Those remain downstream of this pure contract layer.
